+++ src/main/java/domain/slabs/DiscountSlab.java
@@ -3,5 +3,5 @@ package domain.slabs;
 import models.CustomerType;
 
 public interface DiscountSlab {
-    long discount(CustomerType customerType, long purchaseAmount);
+    double discount(CustomerType customerType, double purchaseAmount);
 }
